added stuff

+++ app/actionts.ts
@@ -2,13 +2,20 @@
 
 import { revalidatePath } from "next/cache";
 import prisma from "./lib/db";
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
+import { redirect } from "next/navigation";
 
 export async function createMessage(formData: FormData) {
+  const { getUser } = getKindeServerSession();
+  const user = await getUser();
+
+  if (!user) {
+    return redirect("/api/auth/register");
+  }
   const data = await prisma.guestbookEntry.create({
     data: {
       message: formData.get("message") as string,
-      //get user id
-      userId: "adf",
+      userId: user.id,
     },
   });
 
@@ -16,6 +23,12 @@ export async function createMessage(formData: FormData) {
 }
 
 export async function deleteMessage(formData: FormData) {
+  const { getUser, getPermission } = getKindeServerSession();
+  const user = await getUser();
+
+  if (!user) {
+    return redirect("/api/auth/register");
+  }
   const data = await prisma.guestbookEntry.delete({
     where: {
       id: formData.get("messageId") as string,
